Penyesuaian

+++ database/migrations/2026_05_08_191924_remove_book_id_from_bookrent_table.php
@@ -3,29 +3,29 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('bookrent', function (Blueprint $table) {
-
-            // hapus foreign key dulu
-            $table->dropForeign(['book_id']);
-
-            // hapus kolom
-            $table->dropColumn('book_id');
-
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('bookrent', function (Blueprint $table) {
+                // hapus foreign key dulu
+                $table->dropForeign(['book_id']);
+                // hapus kolom
+                $table->dropColumn('book_id');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('bookrent', function (Blueprint $table) {
-
-            $table->foreignId('book_id')
-                  ->constrained('books');
-
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('bookrent', function (Blueprint $table) {
+                $table->foreignId('book_id')
+                      ->constrained('books');
+            });
+        }
     }
 };
